+++ apps/vs-code-designer/src/app/commands/dataMapper/DataMapperPanel.ts
@@ -92,7 +92,7 @@ export default class DataMapperPanel {
 
   private watchFolderForChanges(folderPath: string, fileExtensions: string[], fn: () => void) {
     // Watch folder for changes to update available file list within Data Mapper
-    const absoluteFolderPath = path.join(ext.logicAppWorkspace, folderPath);
+    const absoluteFolderPath = path.join(ext.defaultLogicAppPath, folderPath);
     if (fileExistsSync(absoluteFolderPath)) {
       const folderWatcher = workspace.createFileSystemWatcher(new RelativePattern(absoluteFolderPath, `**/*.{${fileExtensions.join()}}`));
       folderWatcher.onDidCreate(fn);
@@ -103,10 +103,10 @@ export default class DataMapperPanel {
   }
 
   private setCustomFolders() {
-    const customXsltFullPath = path.join(ext.logicAppWorkspace, customXsltPath);
+    const customXsltFullPath = path.join(ext.defaultLogicAppPath, customXsltPath);
     mkdirSync(customXsltFullPath, { recursive: true });
 
-    const customFunctionsFullPath = path.join(ext.logicAppWorkspace, customFunctionsPath);
+    const customFunctionsFullPath = path.join(ext.defaultLogicAppPath, customFunctionsPath);
     mkdirSync(customFunctionsFullPath, { recursive: true });
   }
 
@@ -133,7 +133,7 @@ export default class DataMapperPanel {
         // Send runtime port to webview
         this.sendMsgToWebview({
           command: ExtensionCommand.setRuntimePort,
-          data: `${ext.designTimePort}`,
+          data: `${ext.designTimeInstances.get(ext.defaultLogicAppPath)?.port}`,
         });
 
         // If loading a data map, handle that + xslt filename
@@ -237,7 +237,7 @@ export default class DataMapperPanel {
   }
 
   public getNestedFilePaths(fileName: string, parentPath: string, relativePath: string, filesToDisplay: string[], filetypes: string[]) {
-    const rootPath = path.join(ext.logicAppWorkspace, relativePath);
+    const rootPath = path.join(ext.defaultLogicAppPath, relativePath);
     const absolutePath = path.join(rootPath, parentPath, fileName);
     if (statSync(absolutePath).isDirectory()) {
       readdirSync(absolutePath).forEach((childFileName) => {
@@ -260,7 +260,7 @@ export default class DataMapperPanel {
     filesToDisplay: IFileSysTreeItem[],
     filetypes: string[]
   ) {
-    const rootPath = path.join(ext.logicAppWorkspace, relativePath);
+    const rootPath = path.join(ext.defaultLogicAppPath, relativePath);
     const absolutePath = path.join(rootPath, parentPath, fileName);
     if (statSync(absolutePath).isDirectory()) {
       const childrenFilesToDisplay: IFileSysTreeItem[] = [];
@@ -299,7 +299,7 @@ export default class DataMapperPanel {
     if (this.dataMapVersion === 2) {
       return this.getFilesTreeForPath(customXsltPath, supportedCustomXsltFileExts, ExtensionCommand.getAvailableCustomXsltPathsV2);
     }
-    const absoluteFolderPath = path.join(ext.logicAppWorkspace, customXsltPath);
+    const absoluteFolderPath = path.join(ext.defaultLogicAppPath, customXsltPath);
     if (fileExistsSync(absoluteFolderPath)) {
       return this.getFilesForPath(customXsltPath, ExtensionCommand.getAvailableCustomXsltPaths, supportedCustomXsltFileExts);
     }
@@ -332,7 +332,7 @@ export default class DataMapperPanel {
     command: typeof ExtensionCommand.showAvailableSchemas | typeof ExtensionCommand.getAvailableCustomXsltPaths,
     fileTypes: string[]
   ) {
-    fs.readdir(path.join(ext.logicAppWorkspace, folderPath)).then((result) => {
+    fs.readdir(path.join(ext.defaultLogicAppPath, folderPath)).then((result) => {
       const filesToDisplay: string[] = [];
       result.forEach((file) => {
         this.getNestedFilePaths(file, '', folderPath, filesToDisplay, fileTypes);
@@ -349,7 +349,7 @@ export default class DataMapperPanel {
     fileTypes: string[],
     command: typeof ExtensionCommand.showAvailableSchemasV2 | typeof ExtensionCommand.getAvailableCustomXsltPathsV2
   ) {
-    fs.readdir(path.join(ext.logicAppWorkspace, folderPath)).then((result) => {
+    fs.readdir(path.join(ext.defaultLogicAppPath, folderPath)).then((result) => {
       const filesToDisplay: IFileSysTreeItem[] = [];
       result.forEach((file) => {
         this.getNestedFileTreePaths(file, '', folderPath, filesToDisplay, fileTypes);
@@ -387,7 +387,7 @@ export default class DataMapperPanel {
         }
         const selectedFile = files[0];
 
-        const pathToWorkspaceSchemaFolder = path.join(ext.logicAppWorkspace, schemasPath);
+        const pathToWorkspaceSchemaFolder = path.join(ext.defaultLogicAppPath, schemasPath);
         const primarySchemaFullPath = selectedFile.fsPath;
         const pathToContainingFolder = path.dirname(primarySchemaFullPath);
         const primarySchemaFileName = path.basename(primarySchemaFullPath);
@@ -421,7 +421,7 @@ export default class DataMapperPanel {
       this.setDataMapperVersionForLogging(context);
 
       const fileName = `${this.dataMapName}${mapDefinitionExtension}`;
-      const dataMapFolderPath = path.join(ext.logicAppWorkspace, dataMapDefinitionsPath);
+      const dataMapFolderPath = path.join(ext.defaultLogicAppPath, dataMapDefinitionsPath);
       const filePath = path.join(dataMapFolderPath, fileName);
 
       // Mkdir as extra insurance that directory exists so file can be written
@@ -453,7 +453,7 @@ export default class DataMapperPanel {
       this.setDataMapperVersionForLogging(context);
 
       const fileName = `${this.dataMapName}${mapXsltExtension}`;
-      const dataMapFolderPath = path.join(ext.logicAppWorkspace, dataMapsPath);
+      const dataMapFolderPath = path.join(ext.defaultLogicAppPath, dataMapsPath);
       const filePath = path.join(dataMapFolderPath, fileName);
 
       // Mkdir as extra insurance that directory exists so file can be written
@@ -477,7 +477,7 @@ export default class DataMapperPanel {
 
   public saveDraftDataMapDefinition(mapDefFileContents: string) {
     const mapDefileName = `${this.dataMapName}${draftMapDefinitionSuffix}${mapDefinitionExtension}`;
-    const dataMapDefFolderPath = path.join(ext.logicAppWorkspace, dataMapDefinitionsPath);
+    const dataMapDefFolderPath = path.join(ext.defaultLogicAppPath, dataMapDefinitionsPath);
     const filePath = path.join(dataMapDefFolderPath, mapDefileName);
 
     // Mkdir as extra insurance that directory exists so file can be written
@@ -530,7 +530,7 @@ export default class DataMapperPanel {
 
   public deleteDraftDataMapDefinition() {
     const draftMapDefinitionPath = path.join(
-      ext.logicAppWorkspace,
+      ext.defaultLogicAppPath,
       dataMapDefinitionsPath,
       `${this.dataMapName}${draftMapDefinitionSuffix}${mapDefinitionExtension}`
     );
@@ -540,7 +540,7 @@ export default class DataMapperPanel {
   }
 
   public checkAndSetXslt() {
-    const expectedXsltPath = path.join(ext.logicAppWorkspace, dataMapsPath, `${this.dataMapName}${mapXsltExtension}`);
+    const expectedXsltPath = path.join(ext.defaultLogicAppPath, dataMapsPath, `${this.dataMapName}${mapXsltExtension}`);
 
     if (fileExistsSync(expectedXsltPath)) {
       fs.readFile(expectedXsltPath, 'utf-8').then((fileContents) => {
@@ -580,7 +580,7 @@ export default class DataMapperPanel {
   }
 
   private getMapMetadataPath() {
-    const projectPath = ext.logicAppWorkspace;
+    const projectPath = ext.defaultLogicAppPath;
     let vscodeFolderPath = '';
     if (this.dataMapVersion === 2) {
       vscodeFolderPath = path.join(projectPath, vscodeFolderName, `${this.dataMapName}DataMapMetadata-v2.json`);